delete topic

+++ src/main/java/org/jharkendar/service/TopicService.java
@@ -1,6 +1,8 @@
 package org.jharkendar.service;
 
 
+import org.jharkendar.data.summary.JpaSummary;
+import org.jharkendar.data.summary.SummaryRepository;
 import org.jharkendar.data.topic.JpaTopic;
 import org.jharkendar.data.topic.TopicRepository;
 import org.jharkendar.rest.topic.PublicTopicDto;
@@ -9,6 +11,7 @@ import org.jharkendar.util.exception.TopicNotFoundException;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.transaction.Transactional;
+import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -19,6 +22,9 @@ public class TopicService {
 
     @Inject
     TopicRepository topicRepository;
+
+    @Inject
+    SummaryRepository summaryRepository;
 
     @Transactional
     public String create(String name) {
@@ -39,7 +45,24 @@ public class TopicService {
     @Transactional
     public void delete(String id) {
         JpaTopic jpaTopic = getByIdInternal(id);
+
+        Collection<JpaSummary> summaries = summaryRepository.findByTopicId(jpaTopic.getId());
+        if (!summaries.isEmpty()) {
+            throw createSummaryExistingException(jpaTopic, summaries);
+        }
+
         topicRepository.delete(jpaTopic);
+    }
+
+    private IllegalStateException createSummaryExistingException(JpaTopic jpaTopic, Collection<JpaSummary> summaries) {
+        return new IllegalStateException(
+                "The following summaries still contain the Topic " +
+                        jpaTopic.getName() +
+                        ": " +
+                        summaries.stream()
+                                .map(JpaSummary::getTitle)
+                                .collect(Collectors.joining(","))
+        );
     }
 
     @Transactional
